# api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "Api Documentation",
    "version": "1.0",
    "title": "Api Documentation",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "host": "localhost:8081",
  "basePath": "/rest",
  "tags": [
    {
      "name": "auth-controller",
      "description": "Auth Controller"
    },
    {
      "name": "measurement-controller",
      "description": "Measurement Controller"
    },
    {
      "name": "user-controller",
      "description": "User Controller"
    }
  ],
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "summary": "Authenticate user and issue token",
        "operationId": "authenticateUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "loginRequestDTO",
            "description": "loginRequestDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/auth/signup": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "summary": "Sign up new user",
        "operationId": "registerUserUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "signUpRequestDTO",
            "description": "signUpRequestDTO",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignUpRequestDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        },
        "deprecated": false
      }
    },
    "/v1/measurement-entries": {
      "post": {
        "tags": [
          "measurement-controller"
        ],
        "summary": "Create new measurement",
        "operationId": "createUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "measurementRequest",
            "description": "measurementRequest",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateMeasurementRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/MeasurementDto"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/v1/measurement-entries/{uid}": {
      "get": {
        "tags": [
          "measurement-controller"
        ],
        "summary": "Sample test endpoint",
        "operationId": "measureUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "uid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/v1/measurement-entries{?page}": {
      "get": {
        "tags": [
          "measurement-controller"
        ],
        "summary": "Get measurement",
        "operationId": "getUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "page",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/MeasurementDto"
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/v1/user/me": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get current user",
        "operationId": "getCurrentUserUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/v1/user/role{?uid}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get user role by uid",
        "operationId": "getRoleByUidUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "description": "uid",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Role"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    },
    "/v1/user{?email}": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Retrieve user by email",
        "operationId": "getUserByEmailUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "deprecated": false
      }
    }
  },
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
    "CreateMeasurementRequest": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "enum": [
            "PULSE",
            "PRESSURE",
            "TEMPERATURE",
            "WEIGHT",
            "GLUCOSE"
          ]
        },
        "value": {
          "type": "string"
        }
      },
      "title": "CreateMeasurementRequest"
    },
    "LoginRequestDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "LoginRequestDto"
    },
    "MeasurementDto": {
      "type": "object",
      "properties": {
        "created": {
          "type": "string",
          "format": "date-time"
        },
        "type": {
          "type": "string",
          "enum": [
            "PULSE",
            "PRESSURE",
            "TEMPERATURE",
            "WEIGHT",
            "GLUCOSE"
          ]
        },
        "uid": {
          "type": "string",
          "format": "uuid"
        },
        "unit": {
          "type": "string",
          "enum": [
            "kg",
            "°C",
            "bmp",
            "mmHG",
            "mmol/L"
          ]
        },
        "value": {
          "type": "string"
        }
      },
      "title": "MeasurementDto"
    },
    "Role": {
      "type": "object",
      "properties": {
        "authority": {
          "type": "string"
        },
        "creationTime": {
          "type": "string",
          "format": "date-time"
        },
        "description": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "uid": {
          "type": "string"
        }
      },
      "title": "Role"
    },
    "SignUpRequestDto": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "SignUpRequestDto"
    },
    "User": {
      "type": "object",
      "properties": {
        "active": {
          "type": "boolean"
        },
        "auth": {
          "type": "string",
          "enum": [
            "BASIC",
            "GOOGLE"
          ]
        },
        "creationTime": {
          "type": "string",
          "format": "date-time"
        },
        "email": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "pictureUrl": {
          "type": "string"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "title": "User"
    }
  }
}